Added logout functionality, email verification Removed back button on user preference checklist
