+++ app/src/main/java/com/example/fithub/ui/profile/ProfileFragment.java
@@ -1,5 +1,6 @@
 package com.example.fithub.ui.profile;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,10 +13,13 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.example.fithub.Login;
 import com.example.fithub.R;
+import com.google.firebase.auth.FirebaseAuth;
 
 public class ProfileFragment extends Fragment {
 
+    FirebaseAuth firebaseAuth;
     private ProfileViewModel profileViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
@@ -24,6 +28,16 @@ public class ProfileFragment extends Fragment {
                 new ViewModelProvider(this).get(ProfileViewModel.class);
         View root = inflater.inflate(R.layout.fragment_profile, container, false);
 
+        View logout = root.findViewById(R.id.logoutBtn);
+
+        logout.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                firebaseAuth.getInstance().signOut();
+                startActivity(new Intent(ProfileFragment.this.getContext(), Login.class));
+            }
+        });
+        
         final TextView textView = new TextView(getContext());
 
         profileViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
@@ -33,5 +47,10 @@ public class ProfileFragment extends Fragment {
             }
         });
         return root;
+    }
+
+    public void logout(View view) {
+        FirebaseAuth.getInstance().signOut();
+        startActivity(new Intent(ProfileFragment.this.getActivity(), Login.class));
     }
 }
