+++ app/src/main/java/com/example/fithub/user_preference_checklist.java
@@ -10,7 +10,6 @@ import android.widget.Button;
 import android.widget.ListView;
 
 public class user_preference_checklist extends AppCompatActivity {
-    Button mBackBtn;
     Button mSaveBtn;
     ListView listviewData;
     ArrayAdapter<String> adapter;
@@ -27,15 +26,7 @@ public class user_preference_checklist extends AppCompatActivity {
                 android.R.layout.simple_list_item_multiple_choice, arrayCatagories);
         listviewData.setAdapter(adapter);
 
-        mBackBtn = findViewById(R.id.backBtn);
         mSaveBtn = findViewById(R.id.saveBtn);
-
-        mBackBtn.setOnClickListener(new View.OnClickListener(){
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(getApplicationContext(),Register.class));
-            }
-        });
 
         mSaveBtn.setOnClickListener(new View.OnClickListener(){
             @Override
